requirements.txt
pdfplumber>=0.11
streamlit>=1.35
pandas>=2.2
plotly>=5.20
python-telegram-bot>=20.0
python-dotenv>=1.0
